oop.library: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@38156 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/oop/getattrbase.c
@@ -61,7 +61,7 @@
     /* Look up ID */
     struct iid_bucket *idb;
     struct HashTable *iidtable = GetOBase(OOPBase)->ob_IIDTable;
-    ULONG base = -1UL;
+    ULONG base = (ULONG)-1;
     
     EnterFunc(bug("OOP_GetAttrBase(interfaceID=%s)\n", interfaceID));
